FIX-0004-packages Move ap-tests from core to ap, make packages for core-tests inside core
