+++ app/src/test/java/ru/easycode/words504/languages/data/cloud/LanguagesServiceCallTest.kt
@@ -1,15 +1,12 @@
-package ru.easycode.words504.languages
+package ru.easycode.words504.languages.data.cloud
 
 import kotlinx.coroutines.runBlocking
 import org.junit.Ignore
 import org.junit.Test
-import ru.easycode.words504.data.cloud.MakeService
-import ru.easycode.words504.languages.data.cloud.LanguagesService
 import ru.easycode.words504.data.cloud.ProvideConverterFactory
 import ru.easycode.words504.data.cloud.ProvideLoggingInterceptor
 import ru.easycode.words504.data.cloud.ProvideOkHttpClientBuilder
 import ru.easycode.words504.data.cloud.ProvideRetrofitBuilder
-import ru.easycode.words504.languages.data.cloud.LanguagesMakeService
 
 // TODO: Remove before the release
 class LanguagesServiceCallTest {
