+++ app/src/test/java/ru/easycode/words504/languages/domain/HandleDomainErrorTest.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.core.domain
+package ru.easycode.words504.languages.domain
 
 import java.net.ConnectException
 import java.net.UnknownHostException
