+++ core/src/test/java/ru/easycode/words504/presentation/SomeViewModelTest.kt
@@ -1,10 +1,9 @@
-package ru.easycode.words504
+package ru.easycode.words504.presentation
 
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
 import org.junit.Test
-import ru.easycode.words504.presentation.BaseViewModel
-import ru.easycode.words504.presentation.DispatchersList
+import ru.easycode.words504.BaseTest
 
 class SomeViewModelTest : BaseTest() {
